restructures some folders
